+++ app/templates/records/archon_detail.html
@@ -35,7 +35,7 @@
 
   {% if record.archon_website %}
     {% call callout(heading='Looking for records?') %}
-      <p>Visit the <a href="{{ record.archon_website | sanitise_record_field | safe }}">{{ pageTitle }} website</a> for information on how to access records held there.</p>
+      <p>Visit the <a href="{{ record.archon_website | normalise_record_field | safe }}">{{ pageTitle }} website</a> for information on how to access records held there.</p>
     {% endcall %}
   {% endif %}
 
@@ -45,7 +45,7 @@
 
         <h2 class="tna-heading-l">Contact information</h2>
 
-        {{ record.description | sanitise_record_field | safe }}
+        {{ record.description | normalise_record_field | safe }}
       </div>
     </div>
 
@@ -54,7 +54,7 @@
         <div class="tna-aside tna-aside--tight tna-background-accent-light full-height-aside">
           <h2 class="tna-heading-l">Information about this archive</h2>
 
-          {{ record.place_description | sanitise_record_field | safe }}
+          {{ record.place_description | normalise_record_field | safe }}
 
         </div>
       </div>
