+++ config/utils/string_processing.py
@@ -1,3 +1,5 @@
+"""Pure string processing helpers with no HTTP or Django dependencies."""
+
 import re
